+++ playwright.config.ts
@@ -17,25 +17,25 @@ export default defineConfig({
     reporter: 'html',
     /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
     use: {
-        /* Base URL to use in actions like `await page.goto('/')`. */
         baseURL: 'http://localhost:3000',
-
-        /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
         trace: 'on-first-retry',
+        viewport: { width: 1280, height: 720 },
     },
 
     /* Configure projects for major browsers */
     projects: [
+        { name: 'setup', testDir: './tests', testMatch: /.*\.setup\.ts/ },
         {
             name: 'chromium',
             use: { ...devices['Desktop Chrome'] },
+            dependencies: ['setup'],
         },
         {
             name: 'api',
             testDir: './tests/api',
             use: {
                 baseURL: 'http://localhost:3000',
-            }
+            },
         },
     ],
 
@@ -43,6 +43,6 @@ export default defineConfig({
     webServer: {
         command: 'pnpm dev',
         url: 'http://localhost:3000',
-        reuseExistingServer: !process.env.CI,
+        reuseExistingServer: true,
     },
 });